Add initial support for 2022.3 platform
