+++ gradle-223.properties
@@ -1,14 +1,14 @@
 # Existent IDE versions can be found in the following repos:
 # https://www.jetbrains.com/intellij-repository/releases/
 # https://www.jetbrains.com/intellij-repository/snapshots/
-ideaVersion=IU-2022.2.2
-clionVersion=CL-2022.2.3
+ideaVersion=IU-223.4884-EAP-CANDIDATE-SNAPSHOT
+clionVersion=CL-223.4884-EAP-CANDIDATE-SNAPSHOT
 
 # https://plugins.jetbrains.com/plugin/12775-native-debugging-support/versions
-nativeDebugPluginVersion=222.4167.21
+nativeDebugPluginVersion=223.4884.65
 # https://plugins.jetbrains.com/plugin/227-psiviewer/versions
-psiViewerPluginVersion=222-SNAPSHOT
+psiViewerPluginVersion=223-SNAPSHOT
 
 # please see https://www.jetbrains.org/intellij/sdk/docs/basics/getting_started/build_number_ranges.html for description
-sinceBuild=222.3048
-untilBuild=222.*
+sinceBuild=223.4884
+untilBuild=231.*
