+++ gradle.properties
@@ -1,5 +1,5 @@
 propertiesPluginEnvironmentNameProperty=platformVersion
-# Supported platforms: 222
+# Supported platforms: 222, 223
 platformVersion=222
 # Supported IDEs: idea, clion
 baseIDE=idea
