+++ src/main/kotlin/org/rust/cargo/project/CargoProjectOpenProcessorBase.kt
@@ -12,13 +12,9 @@ import com.intellij.openapi.vfs.VirtualFile
 import com.intellij.platform.PlatformProjectOpenProcessor
 import com.intellij.projectImport.ProjectOpenProcessor
 import org.rust.cargo.CargoConstants
-import org.rust.cargo.icons.CargoIcons
 import org.rust.cargo.project.model.guessAndSetupRustProject
-import javax.swing.Icon
 
-class CargoProjectOpenProcessor : ProjectOpenProcessor() {
-    override fun getIcon(): Icon = CargoIcons.ICON
-    override fun getName(): String = "Cargo"
+abstract class CargoProjectOpenProcessorBase : ProjectOpenProcessor() {
 
     override fun canOpenProject(file: VirtualFile): Boolean {
         return FileUtil.namesEqual(file.name, CargoConstants.MANIFEST_FILE) ||
